move fetchmostrecentmessage back into previews
